+++ modules/createJestConfig.js
@@ -1,5 +1,4 @@
 /* eslint-disable global-require */
-
 'use strict'
 
 module.exports = (resolve, resolveAppPath) => {