+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Export a `hooks` bag of methods to test react hooks.
+
+### Fixed
+- `jest` and `@testing-library` typings.
+
+### Changed
+- Remove unneeded dependencies
 
 ## [3.1.0] - 2020-03-30